// src/payload/payload-types.ts
export type CardSlider =
  | {
      image: number | Media;
      caption?: string | null;
      id?: string | null;
    }[]
  | null;
// ...
export interface Media {
  id: number;
  alt: string;
  caption?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Property {
  id: number;
  title: string;
  contact?: (number | null) | Agent;
  publishedAt?: string | null;
  hero: {
    propertyType: 'apartment' | 'house' | 'area' | 'commercial';
    advertiseType: 'sell' | 'rent';
    price: number;
    heroImage: number | Media;
    secondaryImage?: number | Media | null;
    priceImage?: number | Media | null;
  };
  content: {
    [k: string]: unknown;
  }[];
  additionalContent?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  slider?: CardSlider;
  slug?: string | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    image?: number | Media | null;
  };
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "agent".
 */
export interface Agent {
  id: number;
  name: string;
  email: string;
  phone: string;
  image: number | Media;
  bio: {
    [k: string]: unknown;
  }[];
  slug?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}